Create MySQL_Books
This is a GUI representation to insert and retrieve data from a table named "book1" in MySQL database. The columns in the table are Serial number(Primary key), Book name and Author name. This script provided a GUI where the user can enter data into the table as well as retrieve data from it by giving the serial number of the book which is the primary key.