Create numpybasics.py
This script uses some functions of numerical python module which is also known as numpy. It takes 2 inputs from the user and makes a 2-dimensional array with all the digits between those numbers.It then asks the user for 2 more numbers and returns only the numbers in the array which are divisible by both these numbers.